Move Ref interface and implementations to powsybl-commons (#3008)
* Move Ref interface and implementations to powsybl-commons
* Add unit tests
