+++ commons/src/main/java/com/powsybl/commons/ref/RefObj.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.iidm.network.impl.util;
+package com.powsybl.commons.ref;
 
 /**
  *
