+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/AbstractAutomationSystem.java
@@ -9,7 +9,7 @@ package com.powsybl.iidm.network.impl;
 
 import com.powsybl.commons.util.trove.TBooleanArrayList;
 import com.powsybl.iidm.network.AutomationSystem;
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 
 import java.util.Objects;
