+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/AbstractIdentifiableAdder.java
@@ -10,7 +10,7 @@ package com.powsybl.iidm.network.impl;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.Identifiable;
 import com.powsybl.iidm.network.Validable;
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 import com.powsybl.iidm.network.util.Identifiables;
 
 import java.util.Arrays;
