+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/GroundImpl.java
@@ -9,7 +9,7 @@ package com.powsybl.iidm.network.impl;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.Ground;
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 
 /**
  * @author Nicolas Rol {@literal <nicolas.rol at rte-france.com>}
