+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/LineAdderImpl.java
@@ -10,7 +10,7 @@ package com.powsybl.iidm.network.impl;
 import com.powsybl.iidm.network.LineAdder;
 import com.powsybl.iidm.network.ValidationException;
 import com.powsybl.iidm.network.ValidationUtil;
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 
 /**
  *
