+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/SubstationImpl.java
@@ -12,7 +12,7 @@ import com.google.common.collect.Lists;
 import com.google.common.primitives.Ints;
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.iidm.network.*;
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 
 import java.util.*;
 import java.util.function.Function;
