+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/SynchronousComponentImpl.java
@@ -9,7 +9,7 @@ package com.powsybl.iidm.network.impl;
 
 import com.powsybl.iidm.network.Network;
 import com.powsybl.iidm.network.components.AbstractSynchronousComponent;
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 
 import java.util.Objects;
