+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/VariantArray.java
@@ -7,7 +7,7 @@
  */
 package com.powsybl.iidm.network.impl;
 
-import com.powsybl.iidm.network.impl.util.Ref;
+import com.powsybl.commons.ref.Ref;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
